add shapes docs content (#1705)
This PR adds a bunch of brain-dump content onto the Shapes docs page.
There's more to add, but I'm submitting what I've done so far.

The main addition is explaining the concept of 'shapes' outside of the
context of creating a custom shape. I think it's good to do the basics
before getting into the world of custom stuff.

Note: I've also moved the Meta section from the Editor page to the
Shapes page.
 
### Change Type

- [x] `documentation` — Changes to the documentation only[^2]

[^1]: publishes a `patch` release, for devDependencies use `internal`
[^2]: will not publish a new version

### Release Notes

- Documentation: Added more info about shapes.

---------

Co-authored-by: Steve Ruiz <steveruizok@gmail.com>